refactor: ファイル名の変更
絶対パス指定に際して使用不可の表現を避けるため
