+++ src/FirebaseInitialize.ts
